add events-file-based regression tests to emissions IT; move emissions input files to matsim-scenarios; some adjustments
