+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/example/RunAverageEmissionToolOfflineExample.java
@@ -21,20 +21,20 @@ package org.matsim.contrib.emissions.example;
 
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.contrib.emissions.EmissionModule;
-import org.matsim.contrib.emissions.EmissionUtils;
 import org.matsim.contrib.emissions.utils.EmissionsConfigGroup;
 import org.matsim.core.api.experimental.events.EventsManager;
 import org.matsim.core.config.Config;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.controler.AbstractModule;
+import org.matsim.core.controler.Controler;
 import org.matsim.core.controler.Injector;
-import org.matsim.core.events.EventsManagerImpl;
+import org.matsim.core.controler.OutputDirectoryHierarchy;
 import org.matsim.core.events.EventsUtils;
 import org.matsim.core.events.MatsimEventsReader;
 import org.matsim.core.events.algorithms.EventWriterXML;
 import org.matsim.core.scenario.ScenarioUtils;
+import org.matsim.core.utils.io.IOUtils;
 import org.matsim.vehicles.MatsimVehicleWriter;
-import org.matsim.vehicles.VehicleUtils;
 
 
 /**
@@ -47,12 +47,12 @@ import org.matsim.vehicles.VehicleUtils;
  */
 public final class RunAverageEmissionToolOfflineExample{
 
-	private static final String configFile = "./scenarios/sampleScenario/testv2_Vehv1/config_average.xml";
+//	private static final String configFile = "./scenarios/sampleScenario/testv2_Vehv1/config_average.xml";
 
-	private static final String eventsFile =  "./scenarios/sampleScenario/5.events.xml.gz";
+	public static final String emissionEventsFilename = "emission.events.offline.xml.gz";
+
 	// (remove dependency of one test/execution path from other. kai/ihab, nov'18)
 
-	private static final String emissionEventOutputFileName = "5.emission.events.offline.xml.gz";
 	private Config config;
 
 	// =======================================================================================================		
@@ -62,13 +62,16 @@ public final class RunAverageEmissionToolOfflineExample{
 		emissionToolOfflineExampleV2.run();
 	}
 
-	public Config prepareConfig() {
-		config = ConfigUtils.loadConfig(configFile, new EmissionsConfigGroup());
-		return config;
-	}
+//	public Config prepareConfig() {
+//		config = ConfigUtils.loadConfig(configFile, new EmissionsConfigGroup());
+//		return config;
+//	}
 
-	public Config prepareConfig(String configFile) {
-		config = ConfigUtils.loadConfig(configFile, new EmissionsConfigGroup());
+	public Config prepareConfig(String args){
+		throw new RuntimeException("execution path no longer exists");
+	}
+	public Config prepareConfig(String [] args) {
+		config = ConfigUtils.loadConfig(args, new EmissionsConfigGroup());
 		EmissionsConfigGroup ecg = ConfigUtils.addOrGetModule( config, EmissionsConfigGroup.class );
 		ecg.setHbefaVehicleDescriptionSource(EmissionsConfigGroup.HbefaVehicleDescriptionSource.fromVehicleTypeDescription);
 		return config;
@@ -76,7 +79,8 @@ public final class RunAverageEmissionToolOfflineExample{
 
 	public void run() {
 		if ( config==null ) {
-			this.prepareConfig() ;
+//			this.prepareConfig() ;
+			throw new RuntimeException( "this execution path no longer exists" );
 		}
 		Scenario scenario = ScenarioUtils.loadScenario(config);
 
@@ -92,20 +96,23 @@ public final class RunAverageEmissionToolOfflineExample{
 				bind( Scenario.class ).toInstance( scenario );
 				bind( EventsManager.class ).toInstance( eventsManager );
 				bind( EmissionModule.class ) ;
+//				bind( OutputDirectoryHierarchy.class );
 			}
 		};;
 
 		com.google.inject.Injector injector = Injector.createInjector(config, module );
 
 		EmissionModule emissionModule = injector.getInstance(EmissionModule.class);
+//		OutputDirectoryHierarchy outputDirectoryHierarchy = injector.getInstance( OutputDirectoryHierarchy.class );
 
 		final String outputDirectory = scenario.getConfig().controler().getOutputDirectory();
-		EventWriterXML emissionEventWriter = new EventWriterXML( outputDirectory + emissionEventOutputFileName );
+		EventWriterXML emissionEventWriter = new EventWriterXML( outputDirectory + emissionEventsFilename );
 		emissionModule.getEmissionEventsManager().addHandler(emissionEventWriter);
 
 		eventsManager.initProcessing();
 		MatsimEventsReader matsimEventsReader = new MatsimEventsReader(eventsManager);
-		matsimEventsReader.readFile(eventsFile);
+//		matsimEventsReader.readFile( "./scenarios/sampleScenario/5.events.xml.gz" );
+		matsimEventsReader.readFile( IOUtils.extendUrl( config.getContext(), "../output_events.xml.gz" ).toString() );
 		eventsManager.finishProcessing();
 
 		emissionEventWriter.closeFile();
