+++ examples/scenarios/emissions-sampleScenario/sample_41_EFA_ColdStart_SubSegm_2020detailed.txt
