+++ examples/scenarios/emissions-sampleScenario/sample_41_EFA_HOT_SubSegm_2020detailed.txt
