+++ examples/scenarios/emissions-sampleScenario/sample_EFA_ColdStart_SubSegm_2005detailed.txt
