+++ examples/scenarios/emissions-sampleScenario/sample_EFA_HOT_SubSegm_2005detailed.txt
